공급업체 회원가입 완료

+++ ezotex/src/main/java/com/ezotex/standard/dto/CompanyDTO.java
@@ -21,7 +21,7 @@ public class CompanyDTO {
 	private MultipartFile companyImgFile;
 	private String section;
 	private int distance;
-	private int requireTime;
+	private Double requireTime;
 	private String addressCode;
 	private String approval;
